Issue #4587: Moved all necessary resource files for NewlineAtEndOfFilCheckTest
