+++ src/test/java/com/puppycrawl/tools/checkstyle/grammars/AstRegressionTest.java
@@ -152,7 +152,7 @@ public class AstRegressionTest extends BaseCheckTestSupport {
     @Test
     public void testNewlineCr() throws Exception {
         verifyAst(super.getPath("/checks/InputNewlineCrAtEndOfFileAst.txt"),
-                super.getPath("/checks/InputNewlineCrAtEndOfFile.java"), true);
+                super.getPath("/checks/InputAstRegressionNewlineCrAtEndOfFile.java"), true);
     }
 
     @Test
