+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/InputNewlineCrAtEndOfFileAst.txt
@@ -25,7 +25,7 @@ INTERFACE_DEF -> INTERFACE_DEF [12:0]
 |   |   `--BLOCK_COMMENT_END -> */ [11:2]
 |   `--LITERAL_PUBLIC -> public [12:0]
 |--LITERAL_INTERFACE -> interface [12:7]
-|--IDENT -> InputNewlineCrAtEndOfFile [12:17]
+|--IDENT -> InputAstRegressionNewlineCrAtEndOfFile [12:17]
 `--OBJBLOCK -> OBJBLOCK [13:0]
     |--LCURLY -> { [13:0]
     |--BLOCK_COMMENT_BEGIN -> /* [14:0]
