+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/newlineatendoffile/InputNewlineLfAtEndOfFile.java
@@ -2,7 +2,7 @@
 // Test case file for checkstyle.
 // Created: 2015
 ////////////////////////////////////////////////////////////////////////////////
-package com.puppycrawl.tools.checkstyle.checks;
+package com.puppycrawl.tools.checkstyle.checks.misc.newlineatendoffile;
 
 /**
  * Test case for detection of an existing LF newline at EOF, using the
